backend/requirements.txt
playwright
fastapi
pymongo
uvicorn
openai 
langgraph
python-dotenv
beautifulsoup4
google_play_scraper